Split up import from the destructive move scripts.
This whole thing was making me pucker when I was running it on
production, even though I know it works fine. This introduced some
duplication, but it also makes it pretty much impossible to accidentally
put a production app in maintenance mode or delete its DB because you
ommitted or included some vars.